+++ src/tasks/batch_integration/metrics/graph_connectivity/script.py
@@ -1,0 +1,35 @@
+import anndata as ad
+import scib
+
+## VIASH START
+par = {
+    'input_integrated': 'resources_test/batch_integration/pancreas/integrated_embedding.h5ad',
+    'output': 'output.h5ad',
+}
+meta = {
+    'functionality_name': 'foo',
+}
+## VIASH END
+
+print('Read input', flush=True)
+adata = ad.read_h5ad(par['input_integrated'])
+
+print('compute score', flush=True)
+score = scib.metrics.graph_connectivity(
+    adata,
+    label_key='label',
+)
+
+print('Create output AnnData object', flush=True)
+output = ad.AnnData(
+    uns={
+        'dataset_id': adata.uns['dataset_id'],
+        'normalization_id': adata.uns['normalization_id'],
+        'method_id': adata.uns['method_id'],
+        'metric_ids': [ meta['functionality_name'] ],
+        'metric_values': [ score ]
+    }
+)
+
+print('Write data to file', flush=True)
+output.write_h5ad(par['output'], compression='gzip')