+++ src/tasks/batch_integration/metrics/kbet/script.py
@@ -1,0 +1,43 @@
+import anndata as ad
+from scib.metrics import kBET
+
+## VIASH START
+par = {
+    'input_integrated': 'resources_test/batch_integration/pancreas/integrated_embedding.h5ad',
+    'output': 'output.h5ad',
+}
+
+meta = {
+    'functionality_name': 'foo',
+}
+## VIASH END
+
+print('Read input', flush=True)
+adata = ad.read_h5ad(par['input_integrated'])
+
+print('compute score', flush=True)
+
+score = kBET(
+    adata,
+    batch_key="batch",
+    label_key="label",
+    type_="embed",
+    embed="X_emb",
+    scaled=True,
+    verbose=False,
+)
+print(score, flush=True)
+
+print('Create output AnnData object', flush=True)
+output = ad.AnnData(
+    uns={
+        'dataset_id': adata.uns['dataset_id'],
+        'normalization_id': adata.uns['normalization_id'],
+        'method_id': adata.uns['method_id'],
+        'metric_ids': [ meta['functionality_name'] ],
+        'metric_values': [ score ]
+    }
+)
+
+print('Write data to file', flush=True)
+output.write_h5ad(par['output'], compression='gzip')